Completing the fetching user details functionality

+++ src/main/java/com/education/education/groups/mappers/GroupsMapper.java
@@ -21,7 +21,7 @@ public class GroupsMapper {
 
   public ChartGroupsForUserRes toChartGroupsForUserRes(
       String groupName,
-      Integer personalGroupActivity) {
+      Long personalGroupActivity) {
     return new ChartGroupsForUserRes(
         groupName,
         personalGroupActivity);
